requirements.txt
Django>=3.2.0
web3>=6.0.0
python-dotenv>=0.19.0
cryptography>=41.0.0
requests>=2.31.0